refactor: inference service becomes more modular
